Clean Emscripten in env/clean.sh (#417)
Add a forgotten removal of Emscripten into the env/clean.sh script. It's
intended to clean all artifacts brought in by env/initialize.sh, but we
forgot to add Emscripten into clean.sh when it was added into
initialize.sh.

This fix is attributed to the WebAssembly build infrastructure, that
is tracked by #177.