Use advanced JS optimizations in Release mode (#420)
Turn on Closure Compiler's advanced optimizations for the compiled
JavaScript code, when the Release mode is on.

This enables things like variable and class renaming and other
nontrivial transformations. Previously we've been using the "simple"
compilation mode, which was only removing extra whitespaces and renaming
local variables, and switching to the advanced compilation was
considered dangerous. But now that we're preparing a big revamp
(WebAssembly) anyway and will do rigorous QA testing, we have a chance
to do this as well.

The size wins are considerable: for the resulting Smart Card Connector
app's CRX file, the delta is: -210KB (-19%) in NaCl builds, and -208KB
(-32%) in WebAssembly builds.