Fix the example library unnecessary rebuilds
This commit fixes the annoying issue that our build scripts were always
rebuilding the JS library in
example_js_standalone_smart_card_client_library.

The root cause turned out to be the location of the "exports.js" source
file. Our build scripts rebuild whenever any *.js file changes or its
parent directory does (which is done in order to catch content changes
and renames), which meant that the "." directory gets included into
consideration. As the "." directory always changes during the make
execution, this was causing the extra rebuild of everything.

The fix is to simply move the exports.js file into a separate
subdirectory.