Emulate pipe() IPC function
Add a fake implementation of the pipe() function, in order to work
around the Emscripten issue with all pipes being created as
non-blocking - which is a non-standard behavior that was causing crashes
inside the PC/SC-Lite daemon.

Also wire up this fake pipe() implementation, together with
corresponding fake write()/read()/close(), with the third-party code in
hotplug_libusb.c, by redefining the standard names to our own helpers.
This is hacky, but allows to not maintain complex patches that'd need to
be reapplied to the upstream.

This commit contributes to the WebAssembly migration effort, tracked by
issue #233.