Rename Detach() to ShutDown()
Rename all Detach() methods in the C++ code to ShutDown(). The new name
is more self-explanatory.

This is a pure refactoring commit. It's done as part of the preparation
for the WebUSB implementation effort, since we'ge going to add a bunch
of new classes and want to use consistent naming.