[wasm] Fix&enable multi-threading in tests
Fix the configuration issues that prevented the usage of
multi-threading in unit tests. This required a few "magic" flags to
Emscripten and tweaking the GoogleTest build parameters, but also we had
to wait until Emscripten fixes a few toolchain issues.

This commit also reenables all multi-threading tests to be run on
Emscripten as well, as they were previously NaCl-only. Given that the
Debug mode is 1-2 orders of magntiude slower in Emscripten than the
Release mode, the commit also tweaks the parameters in stress tests.