[ux] Improve heuristic for spurious USB failures
Improve the heuristic that we use for hiding failed devices from the UI.

The previous one - based on interface numbers - turned out to not work
for some popular devices like YubiKeys, so we ended up displaying two
items for them: a successfully initialized and a failed device, because
it's a composite device for which only the third interface is smart card
related.

The new heuristic is based on display names: if we're trying to show
multiple entries that have the same display name but some are failed and
others succeeded, then the failures are likely spurious.

For history, the last time we touched this heuristic was #362.