[usb] Fix chrome.usb empty output transfer results
Contrary to the chrome.usb API docs, the transfer result's "data" field
is always provided, even for output transfers - for them it's set to an
empty array buffer. Fix our code to stop being confused by that, since
we were misinterpreting this as a zero-length write, effectively the
failure to do output transfers.

This commit is part of the WebUSB support effort tracked by #429.