[usb] Implement transfer timeouts in proxy (#466)
Add implementation of USB transfer timeouts into the libusb-to-JS
proxy. Two main aspects of this are: (1) waiting for the transfer result
shouldn't exceed the timeout, and (2) data that's read after an input
transfer is timed out should be rerouted to later transfers.

The goal is to stop relying on the JavaScript counterpart to enforce the
timeouts, since, unlike chrome.usb, WebUSB doesn't support timeouts.

Luckily, "2" was already implemented by us in the past (for the
cancellation to work), so implementing timeouts in this commit isn't a
radical change.