[usb] Fix output transfer length overflow (#469)
Use 64-bit integers for storing bulk/interrupt transfer length. The
length we used before - 16 bits - is only sufficient for control
transfers, and we incorrectly used it for all types of transfers,
causing the length to be clamped due to integer overflow.

This commit is part of the WebUSB support effort tracked by #429.