[usb] Port loginState hook to new adaptor (#473)
Port ChromeLoginStateHook to the new libusb-to_JS adaptor, so that
Libusb doesn't see any devices when the devices is locked. For this, add
an abstract definition for the hook class and implement the new
loginState hook via it. Thanks to the better libusb-to-JS API and
extensive use of promises, the hook is now implemented in a more
type-check-friendly way.

This commit is part of the WebUSB support effort tracked by #429.