[wasm] Use content security policy in extension
This commit adds a content_security_policy key into manifest.json file
when the EXTENSION packaging mode is used. This allows thw WebAssembly
module to be loaded in that mode, since by default the extension CSP
forbids this.

Note that this is not required in the APP packaging mode, since App's
default CSP is not that strict.