[usb] Compile-time checks for struct field sizes (#475)
Add compile-time assertions into some of the C/C++ code that performs
data conversion to/from libusb structs and to/from libusb-to-JS adaptor
structs. This should prevent bugs with accidental data clamping, similar
to #469.

Also make some of the libusb-to-JS structs have a more precise sizing,
as a few fields were violating the checks added by this commit.

This commit is part of the WebUSB implementation effort, tracked
by #429.