[usb] Implement getConfigurations via WebUSB
Add implementation of the "getConfigurations" command into the
LibusbToWebusbAdaptor JavaScript class.

Thanks to the previous refactoring work, no other changes are needed in
order to let the CCID driver obtain the device configurations via this
WebUSB integration. However, there's one major deficiency: the
"extra_data" field is not set, because WebUSB doesn't expose this
information; backfilling it will be done in a follow-up change.