[usb] Fetch extraData manually when using WebUSB
Polyfill the missing "extraData" information that's not returned by
WebUSB but is part of the libusb interface and is used by the CCID
driver. Do this by making a special "GET_DESCRIPTOR" control transfer to
the device and parsing the bytes out of it.

For the background, each USB device exposes a hierarchy of descriptors
(basically, metadata to be used by the host): configuration descriptors
==> interface descriptors ==> endpoint descriptors. Additionally there
might be class-/vendor-specific descriptors, like 0x21 is a smart card
reader specific descriptor type, and it contains stuff like whether
there's a PIN support in the reader. Such class-/vendor-specific
descriptors aren't returned by WebUSB currently, but the libusb library
that we're trying to emulate does expose it as field "extra_data",
and the CCID driver we're using does heavily rely on this data to be
present.

This polyfill wasn't needed in the past, because the chrome.usb API
returned extraData directly.