[common] Optimize ArrayBuffer debug dumping (#481)
Optimize the JavaScript DebugDump helper that deals with ArrayBuffers
and make the output format more compact (use a single "0x" instead of
prepending "0x" before every byte and using the ", " separator).

The previous implementation was doing lots of overhead by passing every
byte through a bunch of type and boundary checks. All of that is
unnecessary, as we know that after the type casting we're dealing with
an array of unsigned single-byte integers.

We don't have any measurements, but the inefficiency of the old version
could actually be user-visible, since we have a lot of log statements
that are calculated even in the Release mode (unlike when using C/C++,
it's not possible to automatically completely skip log message
calculation in JavaScript, even when we know that we're below the
current log level threshold).