[wasm] Support DataView-to-Value conversion
Add support for conversions from the JavaScript DataView type into the
C++ Value class. The DataView was one of types that we intentionally
skipped from the original emscripten-to-Value converter, because this
type seemed too exotic to be used in our project, but it turned out that
WebUSB uses DataView a lot, so instead of doing artificial conversions
at the WebUSB adaptor layer we're filling the gap in the
emscripten-to-Value converter.