[usb] Device selector dialog in WebUSB
Add implementation for prompting the user to grant a permission in the
configuration when WebUSB is used.

Due to the WebUSB limitations, we can't show the prompt directly after
the button click in the browser action, therefore we're opening a new
tab with the same window, letting the user to click the button again and
select the device. This is a bad UX, but it's a temporary measure until
Chrome is changed to allowlist the Smart Card Connector for accessing
all smart card readers.