[usb] Implement {open,close}DeviceHandle in WebUSB (#485)
Implement openDeviceHandle()/closeDeviceHandle() requests in the
libusb-to-WebUSB adaptor. As WebUSB doesn't have a concept of device
handles, but Libusb does, our adaptor is generating them using a simple
counter. The main difficulty is to handle the cases of concurrent
open/close requests, as WebUSB doesn't like concurrent open/close calls,
and we want to correctly handle scenarios when an open request comes
before the close finishes.