[usb] Fetch WebUSB extraData via handles (#496)
Switch the polyfill that fetches extraData from direct calls to WebUSB
open()/close() methods to using device handles. This allows to make sure
that we don't try to open a device twice (if a concurrent request is
started later), and that we don't close a device that's still in use
(again by some concurrent operation).

Also while doing this, apply a small optimization: don't open-close the
device for each configuration descriptor, but do this once for all
descriptors of a single device.