[common] Fix error handling in MessageChannelPair
Fix a corner case in the MessageChannelPair which could lead to the
"cannot read property of undefined" exception in some error handling
scenarios.

The problem was that if one channel of the pair gets destroyed quickly
after the message is sent to it, by the time it's delivered (which
happens asynchronously) we were trying to find the message handler using
already cleared data members. An example when this can happen is when
the Emscripten fails to load (e.g., because of a timeout).

The fix has two parts: first, check the disposal state right before
delivering the message, and, second, propagate the disposal of any of
the pair's items to the other item too.