[common] Fix error handling in MessageChannelPair (#503)
Fix a corner case in the MessageChannelPair which could lead to the
"cannot read property of undefined" exception in some error handling
scenarios.

The problem was that if one channel of the pair gets destroyed quickly
after the message is sent to it, by the time it's delivered (which
happens asynchronously) we were trying to find the message handler using
already cleared data members. An example when this can happen is when
the Emscripten module fails to load (repro example: set a breakpoint in
DevTools shortly after we start loading the module, wait a minute after the
breakpoint is hit and then continue), and the exception would be triggered
while trying to deliver a PC/SC command message to the module.

The fix has two parts: first, check the disposal state right before
delivering the message, and, second, propagate the disposal of any of
the pair's items to the other item too.