[js] Fix unit tests for optimized mode (#517)
Fix a few tests to make them work in the advanced optimizations mode of
Closure Compiler. This paves the way for enabling this mode in unit
tests in Release builds - currently tests are built without
optimizations, which puts us under risk of shipping broken code.

The challenge with unit tests is that many mocking tricks are broken
when the compiler renames fields and methods, inlines function calls
and substitutes constants. Hence in this commit we go back to classic
methods like "overrideXForTesting" and class inheritance in a few
places - these approaches are guaranteed to work with any kind of
optimizations.

Other issues fixed in this commit are mostly related to quoting object
properties where needed, in case the test injects or expects data in
particular format.