[allowlist] Autoapprove permission for kiosk apps (#526)
Add a few kiosk apps that use smart card communication. Autoapprove
their permissions (without asking the user), in order to mitigate the
issue with invisible prompts in the kiosk mode when the Smart Card
Connector runs as a secondary app (normally the solution to this
problem is policy-for-extensions, however it doesn't exist for
secondary kiosk apps).