Install Python 2 locally under /env/
Stop relying on the system-wide Python 2 installation. Due to the sunset
of Python 2, it's becoming increasingly cumbersome to have it available
and not conflict with Python 3: both on developer machines and on
Continuous Integration bots.

So instead we update the env/initialize.sh script to bootstrap a
completely independent Python 2 installation (compiled from sources and
installed under ./env/python2/). We also create an associated
virtualenv, so that it's easy to temporarily switch into the "python is
Python 2" mode and have the needed packages installed there.

A final note: all this will become unnecessary once we drop NaCl
support (in some mid-term future). However, we can't wait until this
happens and suffer from lacking development processes and CI.