Refactor SingleMessageBasedChannel JS tests
Refactor the unit tests of the SingleMessageBasedChannel JavaScript
class to be more readable and compact, by making a heavier use of
async/await and by extracting repeated blocks into SetUp/TearDown
methods.

This also fixes the flakiness that was caused by the tests not being
switched to the new override{Timeout|Interval}ForTesting() methods.