Bring back support of extension IDs in policies
Fix the regression introduced in the effort to switch to generic
messaging origins (#377). The regression is that we accidentally
switched the policy syntax from the list of extension IDs to the list of
origins, like "chrome-extension://".

This commit brings the backwards compatibility back, so that we still
recognize the policies that have the list of extension IDs. Using the
origins syntax in the policies is allowed, although not enforced.

This fixes #540.