Mkdir the target dir in Googletest makefiles
Fix a possible flakiness in the Googletest makefile, caused by the fact
it attempts to write into LIB_DIR (env/emsdk/lib/{Debug|Release})
without making sure it's already created. So if Googletest builds happen
before other libraries create LIB_DIR, the compilation will fail.

The fix is to add a dependency on the sentinel "dir.stamp" file that
acts as a declaration to create the parent directories. For consistency,
use this approach not only in the Googletest's makefile, but also in
executable_building_emscripten.mk.