Use 32-bit mode in Coverage builds
Tell Clang to build and link the code in the 32-bit mode in
TOOLCHAIN=coverage builds.

This fixes compilation errors caused by the fact that the "long" type is
the same as "int64_t" in 64-bit builds, which causes some of our
function overloads to become duplicates of each other. Also 32-bit mode
is what NaCl and WebAssembly builds use, hence it's better to use the
same in the coverage builds too.