Stub out main() in Coverage builds
Normal builds (WebAssembly/Emscripten) don't need a main() function,
because the framework takes care of the startup itself. In the coverage
builds, however, we're building using stock Clang, and it fails the
linking stage if there's no main() provided.

Note that the empty main() is just fine for coverage builds, since all
we care about is building an executable that the LLVM instrumentation
can see and understand that it needs to account for all functions that
were present in corresponding object files.