Avoid useless rebuilds in Googletest
Fix the inconvenient behavior that Googletest would be fully recompiled
every time the developer switches between modes (e.g., when compiling in
CONFIG=Debug, then in CONFIG=Release, and the going back to
CONFIG=Debug, we were starting the whole rebuild from scratch). This
issue was caused by the fact that we reused the same temporary build
directory for all toolchains and configs. When switching configs, the
old temporary directory was overwritten, making Make think the installed
libraries are out-of-date too.

The commit fixes this by using a dedicated directory name, like
"our-artifacts-emscripten-Debug".