Upload built app .zips in CI
Extend the Continuous Integration script to upload all built .zip
packages (for the applications built in various build modes) as build
artifacts. This makes it possible for developers to obtain the .zips
from a reliable and consistent environment (Github Actions VMs),
rather than to rely on themselves to compile the code correctly locally.

Additionally flatten the contents of the "built_app_packages" directory
(removing the "Debug"/"Release" subfolders), so that the upload step can
just take individual files from it and upload them as separate
artifacts. Also include the packaging mode ("app"/"extension") into the
.zip name, to avoid any potential confusion.