Don't discard coverage failures in CI
Fix the Continuous Integration script to not swallow errors from the
"coverage-report.sh" script.

If this script failed, we were ignoring this and proceeding to
subsequent steps of the CI script, eventually posting a misleading "All
tests passed" comment on the Pull Request. The reason was that the
coverage step was encoded as a multi-line text, which GitHub interprets
as separate Bash commands, so that only the last command determines the
overall step's exit code. The fix is to use the folded-style YAML text
(">") and add "&&" between commands, so that any failure is bubbled up.