Use relative path to PC/SC-Lite config files
Change the PCSCLITE_HP_DROPDIR constant to specify a relative file path
(without a leading slash).

This is needed in order to be able to run the PC/SC-Lite code in unit
tests, where, unlike running without a Chrome Extension/App, we can't
create a virtual file system with this folder in root. We also don't
want to conditionally change this constant depending on test/non-test
environment, hence using this trick with a relative file path.