Add TestingGlobalContext (#594)
Add a test implementation of the GlobalContext class that allows to
mock out the JavaScript counterpart in tests of C++ components. The
class simplifies setting up test expectations on the C++-to-JS messages
and simulating JS-to-C++ replies.

This class will be used by follow-up changes for implementing new unit
tests.