Import PC/SC-Lite ToT .gitignore (#613)
Pull in an upstream change that added a gitignore file into the
PC/SC-Lite repository. We don't create a "patch" file for this
because we'll anyway pull the rest of the PC/SC-Lite's updated
sources soon.

This gitignore fiole will be used by follow-up changes (#612) in
order to clean up build artifacts created by the original
PC/SC-Lite build scripts.