Run PC/SC-Lite's autoconf scripts in our builds (#612)
Change our makefiles to also run PC/SC-Lite's original "bootstrap" and
"configure" scripts (which are based on autoconf themselves).

This will be needed as a follow-up change is about to uprev PC/SC-Lite
to 1.9.7, and that version stopped distributing some autogenerated
headers (pcsclite.h and pcscd.h). We could try replicating the
autogeneration rules in our own makefiles, but a more future-proof
solution seems to just let the original PC/SC-Lite's build scripts
prepare them - which is what this commit implements.

Note that, as a side effect, we start requiring new dependencies that
a developer has to install on their machines: "autoconf",
"autoconf-archive" and "pkg-config".