Fix PC/SC-Lite build_configuration file removing
Fix a bug in the pcsc-lite/naclport/build_configuration makefile that
was causing unrelated files to be removed. The issue was that we were
piping the output of "git status --ignored" to "rm", however git status
emits output in a human-readable format that mentions other files as
well (and, generally speaking, other words that can accidentally
coincide with real file names).

Instead, just use "git clean" to do the right thing.