Fix PC/SC-Lite build_configuration file removing (#619)
Fix a bug in the pcsc-lite/naclport/build_configuration makefile that
was causing unrelated files to be removed. The issue was that we were
piping the output of "git status --ignored" to "rm", however git status
emits output in a human-readable format that mentions other files as
well (and, generally speaking, other words that can accidentally
coincide with real file names). We ended up accidentally deleting
unrelated files when running "make".

This commit fixes it, by simply using "git clean" to do the right thing.