Use relative path for fake_socket_file
Update one more global file path to a relative file path: the
FAKE_PCSC_NACL_SOCKET_FILE_NAME constant that point to the
"fake_socket_file" (for the background, the only purpose of this file is
to mimick the presence of the Unix socket file that PC/SC-Lite normally
uses for IPC; we mock out the actual IPC, but the file presence checks
can't be mocked out without patching the PC/SC-Lite sources).

Using relative paths allows to run the same code in unit tests in
"coverage" mode, which doesn't have a virtual file system for exposing
files under absolute paths in "/".