Backport PC/SC-Lite upstream shutdown fixes
Pull several fixes that we committed into the upstream PC/SC-Lite
project. These make it possible to have unit tests that exercise the
PC/SC-Lite daemon code without leaking or corrupting memory and without
one test affecting the results of subsequent ones.

We also commit the .patch files in order to track the changes we're
doing with original 1.9.7 sources locally. The .patch files can be
deleted after the upstream releases a new version and we import it.