Refactor resource file packaging build rules (#630)
Add a common makefile macro for adding a static resource file into the
resulting executable module file system. The macro has different
implementations depending on the toolchain (e.g., for Emscripten we
tell linker to "preload" the files into a .data archive, and for
Coverage we just copy the file).

The benefit is simplifying the smart_card_connector_app's Makefile by
removing toolchain-specific details from it. Additionally, it allows us
to add resource files to unit test binaries without much hassle (it'll
be done in a follow-up change).