Workaround PC/SC-Lite shutdown timing issues (#639)
Increase the sleep timeout in the PC/SC-Lite shutdown code that we
copied from upstream. This should make it much less likely that we exit
the shutdown function before the background "hotplug_libusb" thread
actually finishes, which was causing flakiness with some probability.

A proper fix will need to be found later.