Don't sleep before abort() in coverage builds (#641)
Crash immediately when some assertion fails, instead of sleeping
for 1 second, in TOOLCHAIN=coverage builds. This makes it much
simpler to investigate test failures.

We continue to use sleeping in WebAssembly/NaCl builds, because there it
increases the chances that logs of the event that led to the crash reach
the JavaScript side.