Run CI in background once a day (#660)
Schedule an additional invocation of Continuous Integration every day at
midnight. The first goal is to increase chances of catching test
flakiness in the landed code. The second goal is to detect any
regressions in the CI script (e.g., due to GitHub's Ubuntu uprev) before
the next person will upload a PR.