Assert on unexpected USB transfers in unit tests
Make the Smart Card Connector C++ unit tests assert when seeing an
unexpected USB request, instead of replying with an USB error.

This simplifies investigating test failures and development of new ones,
since our goal is to be prepared for everything that PC/SC&CCID can send
in tests, and when there's some gap it's hard to understand which USB
transfer was the culprit.