Unit-test Connector unknown PC/SC function call
Add unit test for the Smart Card Connector for the scenario in which it
receives a PC/SC request with an unknown function name.

This change includes the refactoring that was needed in order to
properly parse failure response messages (previously the tests were
parsing both requests and response messages in the same way, assuming
that the "payload" field is always present).