Fix not accounting LOG_DEBUG in coverage builds
This commit makes sure the GOOGLE_SMART_CARD_LOG_DEBUG statements are
treated as covered in the coverage reports.

Before this commit, this wasn't the case, because the macro expanded to
different code in CONFIG=Debug and CONFIG=Release builds (which is
intentional, so that unnecessary debug dumping is stripped away in
release builds). So we make a coverage-build-only adaptation that
suppresses this behavior and hence leads to correct coverage figures.