Fix not accounting LOG_DEBUG in coverage builds (#679)
This commit makes sure the GOOGLE_SMART_CARD_LOG_DEBUG
statements are treated as covered in the coverage reports.

Before this commit, this wasn't the case, because the macro expanded to
different code in CONFIG=Debug and CONFIG=Release builds (which is
intentional, so that unnecessary debug dumping is stripped away in
release builds). LLVM-COV assumes these two different expansions need to
be accounted for differently, and shows two different versions of every code
block that uses LOG_DEBUG: one for the Debug build (which is shown as
covered), and another for the Release build (shown uncovered).

So we make a coverage-build-only adaptation that suppresses the log
statements optimization, so that the LOG_DEBUG macro expands to the
same source code and the coverage figures are correct.